Moved trig functions to Horner's method for evaluating the Maclaurin polynomials for sine, using sine to calculate the cosine and double precision throughout

+++ src/Maths.h
@@ -5,8 +5,8 @@ constexpr float PI = 3.14159265359f;
 float floor(float x);
 float ceil(float x);
 
-float sin(float x);
-float cos(float x);
+double sin(double x);
+double cos(double x);
 
 float fmod(float x, float y);
 bool ispositive(float x);
